add extension for new configuration API
