+++ src/NServiceBus.Gateway.Tests/ApprovalFiles/APIApprovals.Approve.netstandard.approved.txt
@@ -87,6 +87,7 @@ namespace NServiceBus
     public class static GatewayExtensions
     {
         public static NServiceBus.GatewaySettings Gateway(this NServiceBus.EndpointConfiguration config) { }
+        public static NServiceBus.GatewaySettings Gateway(this NServiceBus.EndpointConfiguration config, NServiceBus.Gateway.IGatewayDeduplicationConfiguration storageConfiguration) { }
     }
     public class GatewaySettings
     {
