+++ src/NServiceBus.Gateway/Deduplication/LegacyDeduplicationStorageConfiguration.cs
@@ -1,0 +1,13 @@
+﻿namespace NServiceBus.Gateway
+{
+    using Deduplication;
+    using ObjectBuilder;
+
+    class LegacyDeduplicationStorageConfiguration : IGatewayDeduplicationConfiguration
+    {
+        public IGatewayDeduplicationStorage CreateStorage(IBuilder builder)
+        {
+            return new LegacyDeduplicationWrapper(builder.Build<IDeduplicateMessages>());
+        }
+    }
+}
